Move processing module interface specific defines to separate file.
These defines need to be included in C++ wrappers for
ProcesingModuleInterface defined in Intel IADK based
3rd party modules. Move it to separate file to not include
whole sof stuff there.
